+++ src/main/java/org/hishatakaran/backend/repository/SettlementRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SettlementRepository extends JpaRepository<Settlement, Integer> {
-    List<Settlement> findByRegionId(Integer regionId);
+public interface SettlementRepository extends JpaRepository<Settlement, Long> {
+    List<Settlement> findByRegionId(Long regionId);
 }
